final_project/requirements.txt
streamlit
google-genai
python-dotenv
behave
selenium